Renamed DateRangeComponentContainer.js styled components file
