+++ src/components/styled-components/__tests__/DateRangeComponentContainers.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { shallow } from "enzyme";
-import { DateRangeSliderContainer } from "../DateRangeSliderContainer";
+import { DateRangeSliderContainer } from "../DateRangeComponentContainer";
 
 test("render DateRangeSliderContainer", () => {
   shallow(<DateRangeSliderContainer />);
